Modified the DataSink class and DataSinkInputSpec class to be able to handle uploading data to S3 by including "s3://bucket_name/.." in the base_directory, passes all unittests in https://github.com/FCP-INDI/C-PAC/blob/test_dev/test/unit/nipype/datasink_test.py

+++ nipype/interfaces/io.py
@@ -131,31 +131,85 @@ class IOBase(BaseInterface):
         return base
 
 
+# Class to track percentage of S3 file upload
+class ProgressPercentage(object):
+    '''
+    Call-able class instsance (via __call__ method) that displays
+    upload percentage of a file to S3
+    '''
+
+    def __init__(self, filename):
+        '''
+        '''
+
+        # Import packages
+        import threading
+        import os
+
+        # Initialize data attributes
+        self._filename = filename
+        self._size = float(os.path.getsize(filename))
+        self._seen_so_far = 0
+        self._lock = threading.Lock()
+
+    def __call__(self, bytes_amount):
+        '''
+        '''
+
+        # Import packages
+        import sys
+
+        # With the lock on, print upload status
+        with self._lock:
+            self._seen_so_far += bytes_amount
+            percentage = (self._seen_so_far / self._size) * 100
+            progress_str = '%d / %d (%.2f%%)\r'\
+                           % (self._seen_so_far, self._size, percentage)
+
+            # Write to stdout
+            sys.stdout.write(progress_str)
+            sys.stdout.flush()
+
+
+# DataSink inputs
 class DataSinkInputSpec(DynamicTraitedSpec, BaseInterfaceInputSpec):
+    '''
+    '''
+
+    # Init inputspec data attributes
     base_directory = Directory(
         desc='Path to the base directory for storing data.')
     container = traits.Str(
         desc='Folder within base directory in which to store output')
     parameterization = traits.Bool(True, usedefault=True,
                                    desc='store output in parametrized structure')
-    strip_dir = Directory(desc='path to strip out of filename')
+    strip_dir = traits.Directory(desc='path to strip out of filename')
     substitutions = InputMultiPath(traits.Tuple(traits.Str, traits.Str),
                                    desc=('List of 2-tuples reflecting string '
                                          'to substitute and string to replace '
                                          'it with'))
-    regexp_substitutions = InputMultiPath(traits.Tuple(traits.Str, traits.Str),
-                                          desc=('List of 2-tuples reflecting a pair '
-                                                'of a Python regexp pattern and a '
-                                                'replacement string. Invoked after '
-                                                'string `substitutions`'))
+    regexp_substitutions = \
+        InputMultiPath(traits.Tuple(traits.Str, traits.Str),
+                       desc=('List of 2-tuples reflecting a pair of a '\
+                             'Python regexp pattern and a replacement '\
+                             'string. Invoked after string `substitutions`'))
 
     _outputs = traits.Dict(traits.Str, value={}, usedefault=True)
     remove_dest_dir = traits.Bool(False, usedefault=True,
                                   desc='remove dest directory when copying dirs')
 
+    # AWS S3 data attributes
+    creds_path = traits.Str(desc='Filepath to AWS credentials file for S3 bucket '\
+                              'access')
+    encrypt_bucket_keys = traits.Bool(desc='Flag indicating whether to use S3 '\
+                                        'server-side AES-256 encryption')
+
+    # Set call-able inputs attributes
     def __setattr__(self, key, value):
+        import nipype.interfaces.traits_extension as nit
+
         if key not in self.copyable_trait_names():
-            if not isdefined(value):
+            if not nit.isdefined(value):
                 super(DataSinkInputSpec, self).__setattr__(key, value)
             self._outputs[key] = value
         else:
@@ -164,11 +218,19 @@ class DataSinkInputSpec(DynamicTraitedSpec, BaseInterfaceInputSpec):
             super(DataSinkInputSpec, self).__setattr__(key, value)
 
 
+# DataSink outputs
 class DataSinkOutputSpec(TraitedSpec):
+    '''
+    '''
 
-    out_file = traits.Any(desc='datasink output')
+    # Import packages
+    import traits.api as tapi
+
+    # Init out file
+    out_file = tapi.Any(desc='datasink output')
 
 
+# Custom DataSink class
 class DataSink(IOBase):
     """ Generic datasink module to store structured outputs
 
@@ -230,9 +292,12 @@ class DataSink(IOBase):
         >>> ds.run() # doctest: +SKIP
 
     """
+
+    # Give obj .inputs and .outputs
     input_spec = DataSinkInputSpec
     output_spec = DataSinkOutputSpec
 
+    # Initialization method to set up datasink
     def __init__(self, infields=None, force_run=True, **kwargs):
         """
         Parameters
@@ -254,6 +319,7 @@ class DataSink(IOBase):
         if force_run:
             self._always_run = True
 
+    # Get destination paths
     def _get_dst(self, src):
         # If path is directory with trailing os.path.sep,
         # then remove that for a more robust behavior
@@ -277,6 +343,7 @@ class DataSink(IOBase):
             dst = dst[1:]
         return dst
 
+    # Substitute paths in substitutions dictionary parameter
     def _substitute(self, pathstr):
         pathstr_ = pathstr
         if isdefined(self.inputs.substitutions):
@@ -297,17 +364,251 @@ class DataSink(IOBase):
             iflogger.info('sub: %s -> %s' % (pathstr_, pathstr))
         return pathstr
 
+    # Check for s3 in base directory
+    def _check_s3_base_dir(self):
+        '''
+        Method to see if the datasink's base directory specifies an
+        S3 bucket path; it it does, it parses the path for the bucket
+        name in the form 's3://bucket_name/...' and adds a bucket
+        attribute to the data sink instance, i.e. self.bucket
+
+        Parameters
+        ----------
+
+        Returns
+        -------
+        s3_flag : boolean
+            flag indicating whether the base_directory contained an
+            S3 bucket path
+        '''
+
+        # Import packages
+        import os
+        import sys
+
+        # Init variables
+        s3_str = 's3://'
+        sep = os.path.sep
+        base_directory = self.inputs.base_directory
+
+        # Check if 's3://' in base dir
+        if base_directory.startswith(s3_str):
+            try:
+                # Expects bucket name to be 's3://bucket_name/base_dir/..'
+                bucket_name = base_directory.split(s3_str)[1].split(sep)[0]
+                # Get the actual bucket object
+                self.bucket = self._fetch_bucket(bucket_name)
+            except Exception as exc:
+                err_msg = 'Unable to access S3 bucket. Error:\n%s. Exiting...'\
+                          % exc
+                print err_msg
+                sys.exit()
+            # Bucket access was a success, set flag
+            s3_flag = True
+        # Otherwise it's just a normal datasink
+        else:
+            s3_flag = False
+
+        # Return s3_flag
+        return s3_flag
+
+    # Function to return AWS secure environment variables
+    def _return_aws_keys(self, creds_path):
+        '''
+        Method to return AWS access key id and secret access key using
+        credentials found in a local file.
+
+        Parameters
+        ----------
+        creds_path : string (filepath)
+            path to the csv file with 'AWSAccessKeyId=' followed by access
+            key in the first row and 'AWSSecretAccessKey=' followed by
+            secret access key in the second row
+
+        Returns
+        -------
+        aws_access_key_id : string
+            string of the AWS access key ID
+        aws_secret_access_key : string
+            string of the AWS secret access key
+        '''
+
+        # Import packages
+        import csv
+
+        # Init variables
+        csv_reader = csv.reader(open(creds_path, 'r'))
+
+        # Grab csv rows
+        row1 = csv_reader.next()[0]
+        row2 = csv_reader.next()[0]
+
+        # And split out for keys
+        aws_access_key_id = row1.split('=')[1]
+        aws_secret_access_key = row2.split('=')[1]
+
+        # Return keys
+        return aws_access_key_id, aws_secret_access_key
+
+    # Fetch bucket object
+    def _fetch_bucket(self, bucket_name):
+        '''
+        Method to a return a bucket object which can be used to interact
+        with an AWS S3 bucket using credentials found in a local file.
+
+        Parameters
+        ----------
+        bucket_name : string
+            string corresponding to the name of the bucket on S3
+
+        Returns
+        -------
+        bucket : boto3.resources.factory.s3.Bucket
+            boto3 s3 Bucket object which is used to interact with files
+            in an S3 bucket on AWS
+        '''
+
+        # Import packages
+        import logging
+
+        try:
+            import boto3
+            import botocore
+        except ImportError as exc:
+            err_msg = 'Boto3 package is not installed - install boto3 and '\
+                      'try again.'
+            raise Exception(err_msg)
+
+        # Init variables
+        creds_path = self.inputs.creds_path
+        iflogger = logging.getLogger('interface')
+
+        # Try and get AWS credentials if a creds_path is specified
+        if creds_path:
+            try:
+                aws_access_key_id, aws_secret_access_key = \
+                    self._return_aws_keys(creds_path)
+            except Exception as exc:
+                err_msg = 'There was a problem extracting the AWS credentials '\
+                          'from the credentials file provided: %s. Error:\n%s'\
+                          % (creds_path, exc)
+                raise Exception(err_msg)
+            # Init connection
+            iflogger.info('Connecting to S3 bucket: %s with credentials from '\
+                          '%s ...' % (bucket_name, creds_path))
+            # Use individual session for each instance of DataSink
+            # Better when datasinks are being used in multi-threading, see:
+            # http://boto3.readthedocs.org/en/latest/guide/resources.html#multithreading
+            session = boto3.session.Session(aws_access_key_id=aws_access_key_id,
+                                            aws_secret_access_key=aws_secret_access_key)
+            s3_resource = session.resource('s3', use_ssl=True)
+
+        # Otherwise, connect anonymously
+        else:
+            iflogger.info('Connecting to AWS: %s anonymously...'\
+                          % bucket_name)
+            session = boto3.session.Session()
+            s3_resource = session.resource('s3', use_ssl=True)
+            s3_resource.meta.client.meta.events.register('choose-signer.s3.*',
+                                                         botocore.handlers.disable_signing)
+
+        # Explicitly declare a secure SSL connection for bucket object
+        bucket = s3_resource.Bucket(bucket_name)
+
+        # And try fetch the bucket with the name argument
+        try:
+            s3_resource.meta.client.head_bucket(Bucket=bucket_name)
+        except botocore.exceptions.ClientError as exc:
+            error_code = int(exc.response['Error']['Code'])
+            if error_code == 403:
+                err_msg = 'Access to bucket: %s is denied; check credentials'\
+                          % bucket_name
+                raise Exception(err_msg)
+            elif error_code == 404:
+                err_msg = 'Bucket: %s does not exist; check spelling and try '\
+                          'again' % bucket_name
+                raise Exception(err_msg)
+            else:
+                err_msg = 'Unable to connect to bucket: %s. Error message:\n%s'\
+                          % (bucket_name, exc)
+        except Exception as exc:
+            err_msg = 'Unable to connect to bucket: %s. Error message:\n%s'\
+                      % (bucket_name, exc)
+            raise Exception(err_msg)
+
+        # Return the bucket
+        return bucket
+
+    # Send up to S3 method
+    def _upload_to_s3(self, src, dst):
+        '''
+        Method to upload outputs to S3 bucket instead of on local disk
+        '''
+
+        # Import packages
+        import logging
+        import os
+
+        # Init variables
+        bucket = self.bucket
+        iflogger = logging.getLogger('interface')
+        s3_str = 's3://'
+        s3_prefix = os.path.join(s3_str, bucket.name)
+
+        # If src is a directory, collect files (this assumes dst is a dir too)
+        if os.path.isdir(src):
+            src_files = []
+            for root, dirs, files in os.walk(src):
+                src_files.extend([os.path.join(root, fil) for fil in files])
+            # Make the dst files have the dst folder as base dir
+            dst_files = [os.path.join(dst, src_f.split(src)[1]) \
+                         for src_f in src_files]
+        else:
+            src_files = [src]
+            dst_files = [dst]
+
+        # Iterate over src and copy to dst
+        for src_idx, src_f in enumerate(src_files):
+            # Get destination filename/keyname
+            dst_f = dst_files[src_idx]
+            dst_k = dst_f.replace(s3_prefix, '').lstrip('/')
+
+            # Copy file up to S3 (either encrypted or not)
+            iflogger.info('Copying %s to S3 bucket, %s, as %s...'\
+                          % (src_f, bucket.name, dst_f))
+            if self.inputs.encrypt_bucket_keys:
+                extra_args = {'ServerSideEncryption' : 'AES256'}
+            else:
+                extra_args = {}
+            bucket.upload_file(src_f, dst_k, ExtraArgs=extra_args,
+                               Callback=ProgressPercentage(src_f))
+
+    # List outputs, main run routine
     def _list_outputs(self):
         """Execute this module.
         """
+
+        # Init variables
+        iflogger = logging.getLogger('interface')
         outputs = self.output_spec().get()
         out_files = []
         outdir = self.inputs.base_directory
+        use_hardlink = str2bool(config.get('execution',
+                                               'try_hard_link_datasink'))
+
+        # If base directory isn't given, assume current directory
         if not isdefined(outdir):
             outdir = '.'
-        outdir = os.path.abspath(outdir)
+
+        # Check if base directory reflects S3-bucket upload
+        s3_flag = self._check_s3_base_dir()
+        if not s3_flag:
+            outdir = os.path.abspath(outdir)
+
+        # If container input is given, append that to outdir
         if isdefined(self.inputs.container):
             outdir = os.path.join(outdir, self.inputs.container)
+        # Create the directory if it doesn't exist
         if not os.path.exists(outdir):
             try:
                 os.makedirs(outdir)
@@ -316,8 +617,8 @@ class DataSink(IOBase):
                     pass
                 else:
                     raise(inst)
-        use_hardlink = str2bool(config.get('execution',
-                                           'try_hard_link_datasink') )
+
+        # Iterate through outputs attributes {key : path(s)}
         for key, files in self.inputs._outputs.items():
             if not isdefined(files):
                 continue
@@ -334,44 +635,49 @@ class DataSink(IOBase):
                 if isinstance(files[0], list):
                     files = [item for sublist in files for item in sublist]
 
+            # Iterate through passed-in source files
             for src in filename_to_list(files):
+                # Format src and dst files
                 src = os.path.abspath(src)
-                if os.path.isfile(src):
-                    dst = self._get_dst(src)
-                    dst = os.path.join(tempoutdir, dst)
-                    dst = self._substitute(dst)
-                    path, _ = os.path.split(dst)
-                    if not os.path.exists(path):
-                        try:
-                            os.makedirs(path)
-                        except OSError, inst:
-                            if 'File exists' in inst:
-                                pass
-                            else:
-                                raise(inst)
-                    iflogger.debug("copyfile: %s %s" % (src, dst))
-                    copyfile(src, dst, copy=True, hashmethod='content',
-                             use_hardlink=use_hardlink)
+                if not os.path.isfile(src):
+                    src = os.path.join(src, '')
+                dst = self._get_dst(src)
+                dst = os.path.join(tempoutdir, dst)
+                dst = self._substitute(dst)
+                path, _ = os.path.split(dst)
+
+                # Create output directory if it doesnt exist
+                if not os.path.exists(path):
+                    try:
+                        os.makedirs(path)
+                    except OSError, inst:
+                        if 'File exists' in inst:
+                            pass
+                        else:
+                            raise(inst)
+
+                # If we're uploading to S3
+                if s3_flag:
+                    self._upload_to_s3(src, dst)
                     out_files.append(dst)
-                elif os.path.isdir(src):
-                    dst = self._get_dst(os.path.join(src, ''))
-                    dst = os.path.join(tempoutdir, dst)
-                    dst = self._substitute(dst)
-                    path, _ = os.path.split(dst)
-                    if not os.path.exists(path):
-                        try:
-                            os.makedirs(path)
-                        except OSError, inst:
-                            if 'File exists' in inst:
-                                pass
-                            else:
-                                raise(inst)
-                    if os.path.exists(dst) and self.inputs.remove_dest_dir:
-                        iflogger.debug("removing: %s" % dst)
-                        shutil.rmtree(dst)
-                    iflogger.debug("copydir: %s %s" % (src, dst))
-                    copytree(src, dst)
-                    out_files.append(dst)
+                # Otherwise, copy locally src -> dst
+                else:
+                    # If src is a file, copy it to dst
+                    if os.path.isfile(src):
+                        iflogger.debug('copyfile: %s %s' % (src, dst))
+                        copyfile(src, dst, copy=True, hashmethod='content',
+                                 use_hardlink=use_hardlink)
+                        out_files.append(dst)
+                    # If src is a directory, copy entire contents to dst dir
+                    elif os.path.isdir(src):
+                        if os.path.exists(dst) and self.inputs.remove_dest_dir:
+                            iflogger.debug('removing: %s' % dst)
+                            shutil.rmtree(dst)
+                        iflogger.debug('copydir: %s %s' % (src, dst))
+                        copytree(src, dst)
+                        out_files.append(dst)
+
+        # Return outputs dictionary
         outputs['out_file'] = out_files
 
         return outputs
